Rework scooper fee
Instead of collecting a fixed fee from everyone, we instead have the *protocol* collect a fee for the batch.
This is some base fee (to cover average transaction costs), and an incremental fee (to incentivize larger batches).

However, the total fee is amortized across *all* of the orders; the orders specify a maximum scooper fee that can be collected, but receive a rebate based on the size of the batch they get included in.

For example, imagine the base fee is 1 ADA, and the per-order fee is 0.2 ADA.

A scoop with 1 order would charge 1.2 ada from that one order.

But if we had a scoop with 30 orders, we would collect 7 ADA, but each user would pay only 0.23 ADA!

We make it a definition of an "honest" scooper to include as many orders as possible in a batch (easily detectable when there is a large backlog).

This is implemented by each order specify a max fee they would be willing to pay, and including that ADA in their order.

The scooper then calculates the per-order cost, and includes that in the redeemer. We make sure it doesn't exceed that limit (to prevent just grabbing the max that everyone specifies), and the rest gets distributed back to the user.

Note that we don't enforce that the collected fee is *exactly* the protocol fee, to allow for rounding issues, and we check that the actual protocol fee is at least greater than 0.