Fix a subtle bug
Before this commit, there was a subtle bug: a scooper could include extra orders, and make the input list shorter.

The order script would check that the stake script was present, which it is;
The stake script would check that the pool script was present, which it is;
The pool script would evaluate each order in the input redeemer, each of which is succesfully processed;
The pool script would enforce that each processed input was unique;

But it wouldn't enforce anything about how the *extra* orders are used. So a scooper could use this to steal funds by executing a scoop with just one order in the input list, and steal the rest.

Refer to the note in the code for a really subtle confluence of factors that has to be true for this implementation to be secure.