Refactor the order datum to support 'strategies'
A "strategy" is an order where the exact details aren't yet known,
and will be decided at a later date.  The on-chain order expresses
that "so long as you receive instructions signed by this key, obey
those instructions".

Then, users can run their own agents, or defer to a third party, to
execute those strategies.

For example, someone could place 10k ADA in a "strategy" order,
owned by them and a destination back into the order contract with
a strategy datum; and then run an agent that authorized periodic
dollar-cost-averaging into a token.

Or, this could be used to implement stop losses, trailing stop losses,
take profit triggers, expiring orders, decaying limit orders, etc.

We make it just the order details, rather than the whole datum,
because this offers at least some protection against a compromised actor,
because the destination and owner can't be changed at a later date.