Fix a bug spotted by Philip
A clever scooper could have included a utxo from their own wallet,
with an order datum, to fool the "every order must be processed" counting.
Instead we make sure that every order processed is a script, so that when
we count the scripts, we can compare that to the number of fields in the
indexing set.