Resolve SSW-302
TxPipe pointed out that we had a redundant check for the stake address,
but this revealed a more subtle issue.

Previously, we were checking that the address didn't change between the
pool output and the pool input. Then, we also checked that the staking
key didn't change.

However, this would have locked a pool with the same staking key
forever. The intention is to allow the treasury administrator to change
the staking address if needed.

This resolves it by moving the full address check into the Scoop
redeemer path. In the Withdraw path, we require that the payment key
stay fixed, but allow the staking key to be set to any
allowed_staking_key from the settings datum.

Note that this required us to change the `authorized_staking_keys` to a
list of Credentials, so we could easily compare, though this is expected
to have low impact.