Resolve SSW-303
Switches our recursive implementation of do_2_exp with a heavily
optimized math.pow2 from the aiken standard library.

This version of the module didn't exist at the time, but uses a really
clever trick to optimize. For the first 8 powers of 2, it uses a lookup
into a bytestring to find the result, and only recurses in big leaps
above 2^8.

In our benchmarks, this brings the number of orders per batch from 32 to
35!!