Resolve SSW-301
TxPipe pointed out that the `orders` parameter always equals `[order,
...rest_orders]`; this made the code slightly harder to reason about,
without providing much material benefit.

At one point in time, this was an optimization, since `process_orders`
needed the output and the tail of the list, and we could avoid
destructuring twice. But since then, that logic has been moved into the
singular `process_order` function, and so we can just do the destructure
inside.

Of note, we only destructure on the 4 relevant branches; no need to
destructure on strategy, because we're just calling the process function
recursively to process whatever the actual strategy is.