Resolve SSW-304
The `datum` parameter was redundant, as we were already recursively
passing down each of the fields anyway.

However, as we hyper-optimized the swap operation, we added specific
fields for each swap, and left the other order types unchanged.

By passing in those specific parameters to individual functions, we can
simplify the code and make it easier to reason about during the audit.