Calculate fees more straight-forwardly
Since we have to count the number of orders anyway, we can just do so
upfront; that lets us calculate the amortized protocol fees instead of
relying on the scooper.

We also can just calculate the final fee directly from the base fee, the
number of simple orders, and the number of strategies.

This may save quite a bit of calculation per order, but I haven't run
the benchmark yet.

It's also worth noting that I left amortized_base_fee in the redeemer
for now, so we don't have to update all of our off-chain serialization
code right away, but we should do this before launch.