Remove fees_per_10k from intermediate state
It never changes, so we were just incurring extra cost to construct a
new pool_state with it; instead, we pass it in to the process_orders,
so we can use it to calculate the swap.

I have not benchmarked this, but figured it fit with simplifying the
fees as a response to SSW-305.