Collect protocol fees on donations
I think initially we skipped collecting a fee on donations, because the
user was doing a donation, so it felt weird to charge a fee.

However, the protocol isn't a liquidity provider for every pool, and
needs to be compensated for the work it does to process the donation.

Without this, donations represent a DDOS attack, where the network could
be flooded with cheap donations of a garbage token, and not have to pay
the protocol to process them.