Resolve SSW-101
The settings UTXO required the output to have the exact same value as
the input

But, this means we can't change the size in bytes of the datum, and the
output might be locked if the minUTXO ever increases.

To resolve this, we compare `without_lovelace` of the own_output.value
and own_input.value. This allows the ADA to vary arbitrarily, so the
minUTXO can be adjusted.

Two things to note about the safety of this:
- We never expect an appreciable amount of ADA at this UTXO, essentially
  just the minUTXO, so the fact that someone can take any surplus if the
  minUTXO changes is not an issue
- Only the settings admin and the treasury admin can spend the UTXO
  anyway