Resolve SSW-203
A pool with negative fees would slowly drain its reserves with each
trade; there may be use cases for this, but we haven't thought through
the implications, and so prevent it in this version of the pool script.

A pool with fees greater than 10,000 would never be able to execute, or
would be vulnerable to some other major flaw.

We allow pools with 0% fees and with 100% fees, as they are safe from
the protocol perspective, and someone in the ecosystem may have a
creative use for these values. However, we strongly recommend anyone
building UIs that interface with the protocol display a strong warning
if the pool fee is above some percentage threshold.