Resolve SSW-204
Neither the settings admin or the treasury admin could update the
authorized staking keys, so the list would be locked from the start.

This updates it to allow the treasury admin to update the authorized
staking keys.

This also addresses a readability concern that lead to this bug: we were
just checking that a bunch of fields were unmodified, which makes it
difficult to tell at a glance which fields are allowed to change.

So, instead, we now construct an "allowed" datum, using either the
inputs or outputs as a base, and then overriding the fields that can
change (or should remain the same).