Resolve SSW-307
This makes the MintLP redeemer path more efficient, which will minimize
the impact on scoop costs.

I also took some effort to convince myself in the comments above this
branch that this (and other permutations) were safe, as this is the most
likely (and most damaging) hack of a DEX, is finding some way to mint
extra LP tokens.