Fix subtle pool manage bug
The output_index field doesn't refer to the index within the list of inputs, but the index within the outputs on the tx that produced it.
So this code would have only worked sometimes, coincidentally, and maybe even have introduced subtle exploits.

Instead, we scan over the inputs to find the pool input being spent, so we can make sure the redeemer correctly points to that one.