Implement the basic stableswap algorithm
This implements a "Stableswap" AMM variant developed/popularized by
Curve.

Typically, an AMM uses a constant product curve: as you swap one token
for the other, the price (the slope of the curve) gets steeper. One can
think of this as the pool "holding onto" its remaining tokens so that it
can ensure it always has liquidity to satisfy a trade. Alternatively, if
people are consistently swapping for a specific token, it must be
because that token is more valuable for some reason.

This is a good default if the pool knows nothing about the tokens, as it
ensures that the pool will never run out of liquidity and always be able
to trade one token for another.

However, when you *do* have additional information about the tokens,
such as the tokens being two stablecoins, or tokens highly correlated in
value, this represents an efficiency in the market: users will
experience abnormally high slippage rates, and arbitragers will extract
that value in a follow up trade.

The Curve Stableswap invariant solves this by significantly "flattening"
the curve around the target ratio of tokens: it can trade tokens much
closer to "1 to 1", and transition back to a constant product invariant
at the edges.

Similar to how a constant product pool ensures that x*y = C, a
stableswap ensures that (x*y) + A*(x + y) = D;

A serves as an "amplification" of the linear component, and D represents
the depth of liquidity in the pool.

(note: in practice, we algebraically manipulate the above significantly
so that it can be expressed in more effective units).

Since computing what the new values for this invariant should be is
expensive, we take a more UTxO oriented approach: the batcher provides
the new value in between each order, and we check that the invariant
holds with the new value, and is concordant with the swap they are
trying to perform.

This also removes donations, since they are complicated in the case of
the stableswap invariant.

Still todo:
 - We plan on adding a percentage cut
 - The fee manager should be able to adjust the amplification factors
   within sensible limits
 - All the tests need to be updated
 - Probably plenty more :)