Create a custom pool type for projects
In several cases, projects own 90% of the liquidity in a pool.

These deep liquidity providers have expressed a desire for keeping the
staking rights on the pool, for catalyst, potential airdrops, and
staking rewards.

So, this new pool type is specifically for them:
- There's a new 'staking admin' field, which can change the staking
  address on the pool
- In return, the pool takes a small fee cut to offset the lost staking
  revenue
- There's an "additional scoopers" field, which are keys allowed to
  execute orders against the pool. This enables projects like Butane to
manage their supply and avoid losing value to arbitrage.

Additionally, we might add a "direct swap" feature that doesn't require
the swap to go through an order UTxO.